requirements.txt
# Spock AI RAG System Dependencies
# Core LangChain packages
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0

# Vector store
chromadb>=0.5.0

# Environment and configuration
python-dotenv>=1.0.0

# Document loaders
pypdf>=4.0.0
unstructured>=0.15.0

# Development and testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# =============================================================================
# API Dependencies (spock_api)
# =============================================================================

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Settings management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# SSE/Streaming support (included in FastAPI, but explicit)
sse-starlette>=2.0.0

# HTTP client for testing
httpx>=0.27.0

